# Conflicts:
#	userfiles/modules/admin/modules/market.php
#	userfiles/modules/admin/templates/upload_button.php
#	userfiles/modules/editor/template_settings/index.php
#	userfiles/modules/shop/products/index.php

+++ src/MicroweberPackages/Cache/TaggableFileStore.php
@@ -58,6 +58,12 @@ class TaggableFileStore implements Store
      */
     protected $tags = array();
 
+    /**
+     * Tags that are deleted
+     * @var array
+     */
+    protected static $flushedTags = array();
+
     protected $emitEvents = false;
 
     /**
@@ -126,6 +132,8 @@ class TaggableFileStore implements Store
      */
     public function get($key)
     {
+
+
         //$cacheHandler = new CacheFileHandler();
         $cacheKey = $key . (is_array($this->tags) ? md5(serialize($this->tags)) : false);
 
@@ -219,6 +227,7 @@ class TaggableFileStore implements Store
     {
         //$cacheHandler = new CacheFileHandler();
 
+
         $cacheKey = $key . (is_array($this->tags) ? md5(serialize($this->tags)) : false);
 
         if (isset($this->files->cachedDataMemory[$cacheKey]) and $this->files->cachedDataMemory[$cacheKey] !== $value) {
@@ -227,11 +236,12 @@ class TaggableFileStore implements Store
 
         if (!isset($this->files->cachedDataMemory[$cacheKey])) {
 
+
+            $this->files->cachedDataMemory[$cacheKey] = $value;
             if (!$seconds) {
                 $seconds = now()->addYear(4);
             }
 
-            $this->files->cachedDataMemory[$cacheKey] = $value;
 
             $value = $this->expiration($seconds) . serialize($value);
 
@@ -544,9 +554,18 @@ class TaggableFileStore implements Store
         $findTagPath = $this->_findCachePathByKey($key);
         $findTagPath = $this->getPath() . $findTagPath;
 
-        if ($this->files->exists($findTagPath)) {
-            $this->files->delete($findTagPath);
+        if (!empty($this->tags)) {
+            self::$flushedTags = array_merge(self::$flushedTags, $this->tags);
         }
+
+        try {
+            if ($this->files->exists($findTagPath)) {
+                $this->files->delete($findTagPath);
+            }
+        } catch (\Exception $e) {
+            //
+        }
+
         if ($this->emitEvents) {
             event(new KeyForgotten($key));
         }
@@ -555,6 +574,7 @@ class TaggableFileStore implements Store
 
     public function delete($key)
     {
+
         return $this->forget($key);
     }
 
@@ -567,21 +587,42 @@ class TaggableFileStore implements Store
     {
         $this->files->cachedDataMemory = [];
         if (!empty($this->tags)) {
+
+            if (!$all and $this->isTagFlushed()) {
+                return;
+            }
+
+
             foreach ($this->tags as $tag) {
                 $tagDetails = $this->_getTagMapByName($tag);
                 if (!empty($tagDetails)) {
                     foreach ($tagDetails as $tagDetail) {
                         $tagPath = $this->getPath() . $tagDetail;
-                        if ($this->files->isFile($tagPath)) {
-                            $this->files->delete($tagPath);
+
+                        try {
+                            if ($this->files->isFile($tagPath)) {
+                                $this->files->delete($tagPath);
+                            }
+                        } catch (\Exception $e) {
+                            //
                         }
+
+
                     }
                 }
 
                 $tagMapPath = $this->_getTagMapPathByName($tag);
-                if ($this->files->isFile($tagMapPath)) {
-                    $this->files->delete($tagMapPath);
+
+                try {
+                    if ($this->files->isFile($tagMapPath)) {
+                        $this->files->delete($tagMapPath);
+                    }
+                } catch (\Exception $e) {
+                    //
                 }
+
+                self::$flushedTags = array_merge(self::$flushedTags, $this->tags);
+
                 if (isset($this->tags[$tag])) {
                     unset($this->tags[$tag]);
                 }
@@ -598,8 +639,21 @@ class TaggableFileStore implements Store
                 return false;
             }
 
-            return @$this->files->deleteDirectory($mainCacheDir);
+            try {
+                $this->files->deleteDirectory($mainCacheDir);
+            } catch (\Exception $e) {
+                //
+            }
 
+        }
+    }
+
+    private function isTagFlushed()
+    {
+        if (self::$flushedTags and $this->tags) {
+            if (in_array($this->tags, self::$flushedTags)) {
+                return true;
+            }
         }
     }
